# perf: Average NamedList __getitem__ performance improvement